requirements.txt
python-dotenv
strands-agents
strands-agents-tools
streamlit>=1.28.0
